Fix mocking, remove interface
